`constrain()` should not return `constrained_interval<>`
A `constrained_interval<>` carries the identity of another object. Returning it would allow accidentally referring to the unconstrained variable in a nested constraint.

If this stands the test, some subsequent simplifications might be possible: can we simplify the `any_interval<>` concept to just require a specialization of `interval<>`; and do we need `constrained_interval<>` at all, or would plain lvalue references in the `*_constraint` types be enough?